% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMean.R
\name{find_mean}
\alias{find_mean}
\title{Find mean of columns according to groups and add it to original tibble.}
\usage{
find_mean(tib, col_name, group_col_name)
}
\arguments{
\item{tib}{A tibble. Since the input should be a tibble I named it "tib"}

\item{col_name}{A column name to find mean of. To make it evident I named it "col_name". Function will remove NA from this column}

\item{group_col_name}{Optional parameter. A column name. If provided, data-set will be grouped by this column before finding mean. Since the data is going to be grouped according to this I named it "group_col_name". Function will not remove any NA from this column.}
}
\value{
A tibble
}
\description{
Add a mean column to data-frame after grouping by another column. The new mean column name will add a "mean." prefix to the
original column name ("mean.col_name"). This function uses the mutate \code{\link[dplyr:mutate]{dplyr::mutate()}} function and group_by \code{\link[dplyr:group_by]{dplyr::group_by()}} from the dplyr package.
}
\examples{
find_mean(gapminder::gapminder, lifeExp)
find_mean(gapminder::gapminder, lifeExp, country)
}
